Kernel: Split the ISO9660FileSystem.{cpp,h} files to smaller components
