+++ Kernel/FileSystem/ISO9660FS/Definitions.h
@@ -7,15 +7,7 @@
 #pragma once
 
 #include <AK/EnumBits.h>
-#include <AK/Error.h>
-#include <AK/HashMap.h>
-#include <AK/RecursionDecision.h>
-#include <AK/StringView.h>
 #include <AK/Types.h>
-#include <Kernel/FileSystem/BlockBasedFileSystem.h>
-#include <Kernel/FileSystem/Inode.h>
-#include <Kernel/KBuffer.h>
-#include <Kernel/Library/NonnullLockRefPtr.h>
 
 namespace Kernel {
 
@@ -275,118 +267,4 @@ static_assert(sizeof(VolumePartitionDescriptor) == 2048);
 
 }
 
-class ISO9660Inode;
-class ISO9660DirectoryIterator;
-
-class ISO9660FS final : public BlockBasedFileSystem {
-    friend ISO9660Inode;
-    friend ISO9660DirectoryIterator;
-
-public:
-    struct DirectoryEntry final : public AtomicRefCounted<DirectoryEntry> {
-        u32 extent { 0 };
-        u32 length { 0 };
-
-        // NOTE: This can never be empty if we read the directory successfully.
-        //       We need it as an OwnPtr to default-construct this struct.
-        OwnPtr<KBuffer> blocks;
-
-        static ErrorOr<NonnullLockRefPtr<DirectoryEntry>> try_create(u32 extent, u32 length, OwnPtr<KBuffer> blocks)
-        {
-            return adopt_nonnull_lock_ref_or_enomem(new (nothrow) DirectoryEntry(extent, length, move(blocks)));
-        }
-
-    private:
-        DirectoryEntry(u32 extent, u32 length, OwnPtr<KBuffer> blocks)
-            : extent(extent)
-            , length(length)
-            , blocks(move(blocks))
-        {
-        }
-    };
-
-    static ErrorOr<NonnullLockRefPtr<FileSystem>> try_create(OpenFileDescription&);
-
-    virtual ~ISO9660FS() override;
-    virtual StringView class_name() const override { return "ISO9660FS"sv; }
-    virtual Inode& root_inode() override;
-
-    virtual unsigned total_block_count() const override;
-    virtual unsigned total_inode_count() const override;
-
-    virtual u8 internal_file_type_to_directory_entry_type(DirectoryEntryView const& entry) const override;
-
-    ErrorOr<NonnullLockRefPtr<DirectoryEntry>> directory_entry_for_record(Badge<ISO9660DirectoryIterator>, ISO::DirectoryRecordHeader const* record);
-
-private:
-    ISO9660FS(OpenFileDescription&);
-
-    virtual ErrorOr<void> prepare_to_clear_last_mount() override;
-
-    virtual bool is_initialized_while_locked() override;
-    virtual ErrorOr<void> initialize_while_locked() override;
-
-    ErrorOr<void> parse_volume_set();
-    ErrorOr<void> create_root_inode();
-    ErrorOr<void> calculate_inode_count() const;
-
-    u32 calculate_directory_entry_cache_key(ISO::DirectoryRecordHeader const&);
-
-    ErrorOr<void> visit_directory_record(ISO::DirectoryRecordHeader const& record, Function<ErrorOr<RecursionDecision>(ISO::DirectoryRecordHeader const*)> const& visitor) const;
-
-    OwnPtr<ISO::PrimaryVolumeDescriptor> m_primary_volume;
-    LockRefPtr<ISO9660Inode> m_root_inode;
-
-    mutable u32 m_cached_inode_count { 0 };
-    HashMap<u32, NonnullLockRefPtr<DirectoryEntry>> m_directory_entry_cache;
-};
-
-class ISO9660Inode final : public Inode {
-    friend ISO9660FS;
-
-public:
-    virtual ~ISO9660Inode() override;
-
-    ISO9660FS& fs() { return static_cast<ISO9660FS&>(Inode::fs()); }
-    ISO9660FS const& fs() const { return static_cast<ISO9660FS const&>(Inode::fs()); }
-
-    // ^Inode
-    virtual InodeMetadata metadata() const override;
-    virtual ErrorOr<void> traverse_as_directory(Function<ErrorOr<void>(FileSystem::DirectoryEntryView const&)>) const override;
-    virtual ErrorOr<NonnullLockRefPtr<Inode>> lookup(StringView name) override;
-    virtual ErrorOr<void> flush_metadata() override;
-    virtual ErrorOr<NonnullLockRefPtr<Inode>> create_child(StringView name, mode_t, dev_t, UserID, GroupID) override;
-    virtual ErrorOr<void> add_child(Inode&, StringView name, mode_t) override;
-    virtual ErrorOr<void> remove_child(StringView name) override;
-    virtual ErrorOr<void> chmod(mode_t) override;
-    virtual ErrorOr<void> chown(UserID, GroupID) override;
-    virtual ErrorOr<void> truncate(u64) override;
-    virtual ErrorOr<void> update_timestamps(Optional<time_t> atime, Optional<time_t> ctime, Optional<time_t> mtime) override;
-
-private:
-    // HACK: The base ISO 9660 standard says the maximum filename length is 37
-    // bytes large; however, we can read filenames longer than that right now
-    // without any problems, so let's allow it anyway.
-    static constexpr size_t max_file_identifier_length = 256 - sizeof(ISO::DirectoryRecordHeader);
-
-    // ^Inode
-    virtual ErrorOr<size_t> read_bytes_locked(off_t, size_t, UserOrKernelBuffer& buffer, OpenFileDescription*) const override;
-    virtual ErrorOr<size_t> write_bytes_locked(off_t, size_t, UserOrKernelBuffer const& buffer, OpenFileDescription*) override;
-
-    ISO9660Inode(ISO9660FS&, ISO::DirectoryRecordHeader const& record, StringView name);
-    static ErrorOr<NonnullLockRefPtr<ISO9660Inode>> try_create_from_directory_record(ISO9660FS&, ISO::DirectoryRecordHeader const& record, StringView name);
-
-    static InodeIndex get_inode_index(ISO::DirectoryRecordHeader const& record, StringView name);
-    static StringView get_normalized_filename(ISO::DirectoryRecordHeader const& record, Bytes buffer);
-
-    void create_metadata();
-    time_t parse_numerical_date_time(ISO::NumericalDateAndTime const&);
-
-    InodeMetadata m_metadata;
-    ISO::DirectoryRecordHeader m_record;
-};
-
 }
-
-using Kernel::ISO::has_any_flag;
-using Kernel::ISO::has_flag;
